Refactor: move OpCode and Instruction definition to bytecode.rs
