+++ parser/src/bytecode.rs
@@ -1,0 +1,141 @@
+//! The definition of bytecode data structure that is shared among the bytecode compiler and the interpreter (vm.rs)
+
+use std::io::{Read, Write};
+
+use crate::parser::ReadError;
+
+/// Operational codes for an instruction. Supposed to fit in an u8.
+#[derive(Debug, Clone, Copy)]
+#[repr(u8)]
+pub enum OpCode {
+    LoadLiteral,
+    /// Move values between stack elements, from arg0 to arg1.
+    Move,
+    /// Increment the operand arg0
+    Incr,
+    Add,
+    Sub,
+    Mul,
+    Div,
+    /// Bitwise and (&)
+    BitAnd,
+    /// Bitwise xor (^)
+    BitXor,
+    /// Bitwise or (|)
+    BitOr,
+    /// Logical and (&&)
+    And,
+    /// Logical or (||)
+    Or,
+    /// Logical not (!)
+    Not,
+    /// Bitwise not (~). Interestingly, Rust does not have dedicated bitwise not operator, because
+    /// it has bool type. It can distinguish logical or bitwise operation by the operand type.
+    /// However, we do not have bool type (yet), so we need a dedicated operator for bitwise not, like C.
+    BitNot,
+    /// Get an element of an array (or a table in the future) at arg0 with the key at arg1, and make a copy at arg1.
+    /// Array elements are always Rc wrapped, so the user can assign into it.
+    Get,
+    /// If a value specified with arg0 in the stack is a reference (pointer), dereference it.
+    Deref,
+    /// Compare arg0 and arg1, sets result -1, 0 or 1 to arg0, meaning less, equal and more, respectively
+    // Cmp,
+    Lt,
+    Gt,
+    /// Unconditional jump to arg1.
+    Jmp,
+    /// Conditional jump. If arg0 is truthy, jump to arg1.
+    Jt,
+    /// Conditional jump. If arg0 is falthy, jump to arg1.
+    Jf,
+    /// Call a function with arg0 aruguments on the stack with index arg1.
+    Call,
+    /// Returns from current call stack.
+    Ret,
+    /// Casts a value at arg0 to a type indicated by arg1. I'm feeling this should be a standard library function
+    /// rather than a opcode, but let's finish implementation compatible with AST interpreter first.
+    Cast,
+}
+
+macro_rules! impl_op_from {
+    ($($op:ident),*) => {
+        impl From<u8> for OpCode {
+            #[allow(non_upper_case_globals)]
+            fn from(o: u8) -> Self {
+                $(const $op: u8 = OpCode::$op as u8;)*
+
+                match o {
+                    $($op => Self::$op,)*
+                    _ => panic!("Opcode \"{:02X}\" unrecognized!", o),
+                }
+            }
+        }
+    }
+}
+
+impl_op_from!(
+    LoadLiteral,
+    Move,
+    Incr,
+    Add,
+    Sub,
+    Mul,
+    Div,
+    BitAnd,
+    BitXor,
+    BitOr,
+    And,
+    Or,
+    Not,
+    BitNot,
+    Get,
+    Deref,
+    Lt,
+    Gt,
+    Jmp,
+    Jt,
+    Jf,
+    Call,
+    Ret,
+    Cast
+);
+
+/// A single instruction in a bytecode. OpCodes can have 0 to 2 arguments.
+#[derive(Debug, Clone, Copy)]
+pub struct Instruction {
+    pub(crate) op: OpCode,
+    pub(crate) arg0: u8,
+    pub(crate) arg1: u16,
+}
+
+impl Instruction {
+    pub(crate) fn new(op: OpCode, arg0: u8, arg1: u16) -> Self {
+        Self { op, arg0, arg1 }
+    }
+    pub(crate) fn serialize(&self, writer: &mut impl Write) -> std::io::Result<()> {
+        writer.write_all(&(self.op as u8).to_le_bytes())?;
+        writer.write_all(&self.arg0.to_le_bytes())?;
+        writer.write_all(&self.arg1.to_le_bytes())?;
+        Ok(())
+    }
+
+    pub(crate) fn deserialize(reader: &mut impl Read) -> Result<Self, ReadError> {
+        let mut op = [0u8; std::mem::size_of::<u8>()];
+        reader.read_exact(&mut op)?;
+        let mut arg0 = [0u8; std::mem::size_of::<u8>()];
+        reader.read_exact(&mut arg0)?;
+        let mut arg1 = [0u8; std::mem::size_of::<u16>()];
+        reader.read_exact(&mut arg1)?;
+        Ok(Self {
+            op: u8::from_le_bytes(op).into(),
+            arg0: u8::from_le_bytes(arg0),
+            arg1: u16::from_le_bytes(arg1),
+        })
+    }
+}
+
+impl std::fmt::Display for Instruction {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{:?} {} {}", self.op, self.arg0, self.arg1)
+    }
+}
