+++ parser/src/lib.rs
@@ -1,3 +1,4 @@
+mod bytecode;
 mod compiler;
 mod interpreter;
 mod parser;
@@ -7,6 +8,7 @@ mod type_tags;
 mod value;
 mod vm;
 
+pub use self::bytecode::{Instruction, OpCode};
 pub use self::compiler::*;
 pub use self::interpreter::{coerce_type, run, EvalContext, EvalError, FuncDef};
 pub use self::parser::{span_source as source, ArgDecl, ReadError, Span};
